// src/lib/interfaces/rustplus.ts
export interface Vector4 {
    /**
     * @generated from protobuf field: float x = 1
     */
    x: number;
    /**
     * @generated from protobuf field: float y = 2
     */
    y: number;
    /**
     * @generated from protobuf field: float z = 3
     */
    z: number;
    /**
     * @generated from protobuf field: float w = 4
     */
    w: number;
}
// ...
export interface AppMap {
    /**
     * @generated from protobuf field: uint32 width = 1
     */
    width: number;
    /**
     * @generated from protobuf field: uint32 height = 2
     */
    height: number;
    /**
     * @generated from protobuf field: bytes jpgImage = 3
     */
    jpgImage: Uint8Array;
    /**
     * @generated from protobuf field: int32 oceanMargin = 4
     */
    oceanMargin: number;
    /**
     * @generated from protobuf field: repeated rustplus.AppMap.Monument monuments = 5
     */
    monuments: AppMap_Monument[];
    /**
     * @generated from protobuf field: optional string background = 6
     */
    background?: string;
}
/**
 * @generated from protobuf message rustplus.AppMap.Monument
 */
export interface AppMap_Monument {
    /**
     * @generated from protobuf field: string token = 1
     */
    token: string;
    /**
     * @generated from protobuf field: float x = 2
     */
    x: number;
    /**
     * @generated from protobuf field: float y = 3
     */
    y: number;
}
// ...
export interface AppMarker {
    /**
     * @generated from protobuf field: uint32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: rustplus.AppMarkerType type = 2
     */
    type: AppMarkerType;
    /**
     * @generated from protobuf field: float x = 3
     */
    x: number;
    /**
     * @generated from protobuf field: float y = 4
     */
    y: number;
    /**
     * @generated from protobuf field: uint64 steamId = 5
     */
    steamId: string;
    /**
     * @generated from protobuf field: float rotation = 6
     */
    rotation: number;
    /**
     * @generated from protobuf field: float radius = 7
     */
    radius: number;
    /**
     * @generated from protobuf field: optional rustplus.Vector4 color1 = 8
     */
    color1?: Vector4;
    /**
     * @generated from protobuf field: optional rustplus.Vector4 color2 = 9
     */
    color2?: Vector4;
    /**
     * @generated from protobuf field: float alpha = 10
     */
    alpha: number;
    /**
     * @generated from protobuf field: optional string name = 11
     */
    name?: string;
    /**
     * @generated from protobuf field: bool outOfStock = 12
     */
    outOfStock: boolean;
    /**
     * @generated from protobuf field: repeated rustplus.AppMarker.SellOrder sellOrders = 13
     */
    sellOrders: AppMarker_SellOrder[];
}
/**
 * @generated from protobuf message rustplus.AppMarker.SellOrder
 */
export interface AppMarker_SellOrder {
    /**
     * @generated from protobuf field: int32 itemId = 1
     */
    itemId: number;
    /**
     * @generated from protobuf field: int32 quantity = 2
     */
    quantity: number;
    /**
     * @generated from protobuf field: int32 currencyId = 3
     */
    currencyId: number;
    /**
     * @generated from protobuf field: int32 costPerItem = 4
     */
    costPerItem: number;
    /**
     * @generated from protobuf field: int32 amountInStock = 5
     */
    amountInStock: number;
    /**
     * @generated from protobuf field: bool itemIsBlueprint = 6
     */
    itemIsBlueprint: boolean;
    /**
     * @generated from protobuf field: bool currencyIsBlueprint = 7
     */
    currencyIsBlueprint: boolean;
    /**
     * @generated from protobuf field: float itemCondition = 8
     */
    itemCondition: number;
    /**
     * @generated from protobuf field: float itemConditionMax = 9
     */
    itemConditionMax: number;
    /**
     * @generated from protobuf field: float priceMultiplier = 10
     */
    priceMultiplier: number;
}
/**
 * @generated from protobuf message rustplus.AppMapMarkers
 */
export interface AppMapMarkers {
    /**
     * @generated from protobuf field: repeated rustplus.AppMarker markers = 1
     */
    markers: AppMarker[];
}
// ...
export enum AppMarkerType {
    /**
     * @generated from protobuf enum value: Undefined = 0;
     */
    Undefined = 0,
    /**
     * @generated from protobuf enum value: Player = 1;
     */
    Player = 1,
    /**
     * @generated from protobuf enum value: Explosion = 2;
     */
    Explosion = 2,
    /**
     * @generated from protobuf enum value: VendingMachine = 3;
     */
    VendingMachine = 3,
    /**
     * @generated from protobuf enum value: CH47 = 4;
     */
    CH47 = 4,
    /**
     * @generated from protobuf enum value: CargoShip = 5;
     */
    CargoShip = 5,
    /**
     * @generated from protobuf enum value: Crate = 6;
     */
    Crate = 6,
    /**
     * @generated from protobuf enum value: GenericRadius = 7;
     */
    GenericRadius = 7,
    /**
     * @generated from protobuf enum value: PatrolHelicopter = 8;
     */
    PatrolHelicopter = 8,
    /**
     * @generated from protobuf enum value: TravellingVendor = 9;
     */
    TravellingVendor = 9
}
// ...
export const Vector4 = new Vector4$Type();
// ...
export const AppMap = new AppMap$Type();
// @generated message type with reflection information, may provide speed optimized methods
class AppMap_Monument$Type extends MessageType<AppMap_Monument> {
    constructor() {
        super("rustplus.AppMap.Monument", [
            {
                no: 1,
                name: "token",
                kind: "scalar",
                T: 9 /*ScalarType.STRING*/
            },
            { no: 2, name: "x", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 3, name: "y", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ }
        ]);
    }
    create(value?: PartialMessage<AppMap_Monument>): AppMap_Monument {
        const message = globalThis.Object.create(this.messagePrototype);
        message.token = "";
        message.x = 0;
        message.y = 0;
        if (value !== undefined)
            reflectionMergePartial<AppMap_Monument>(this, message, value);
        return message;
    }
    internalBinaryRead(
        reader: IBinaryReader,
        length: number,
        options: BinaryReadOptions,
        target?: AppMap_Monument
    ): AppMap_Monument {
        const message = target ?? this.create(),
            end = reader.pos + length;
        while (reader.pos < end) {
            const [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string token */ 1:
                    message.token = reader.string();
                    break;
                case /* float x */ 2:
                    message.x = reader.float();
                    break;
                case /* float y */ 3:
                    message.y = reader.float();
                    break;
                default:
                    const u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(
                            `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
                        );
                    const d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(
                            this.typeName,
                            message,
                            fieldNo,
                            wireType,
                            d
                        );
            }
        }
        return message;
    }
    internalBinaryWrite(
        message: AppMap_Monument,
        writer: IBinaryWriter,
        options: BinaryWriteOptions
    ): IBinaryWriter {
        /* string token = 1; */
        if (message.token !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.token);
        /* float x = 2; */
        if (message.x !== 0) writer.tag(2, WireType.Bit32).float(message.x);
        /* float y = 3; */
        if (message.y !== 0) writer.tag(3, WireType.Bit32).float(message.y);
        const u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(
                this.typeName,
                message,
                writer
            );
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message rustplus.AppMap.Monument
 */
export const AppMap_Monument = new AppMap_Monument$Type();
// ...
export const AppMarker = new AppMarker$Type();
// ...
export const AppMarker_SellOrder = new AppMarker_SellOrder$Type();
// @generated message type with reflection information, may provide speed optimized methods
class AppMapMarkers$Type extends MessageType<AppMapMarkers> {
    constructor() {
        super("rustplus.AppMapMarkers", [
            {
                no: 1,
                name: "markers",
                kind: "message",
                repeat: 2 /*RepeatType.UNPACKED*/,
                T: () => AppMarker
            }
        ]);
    }
    create(value?: PartialMessage<AppMapMarkers>): AppMapMarkers {
        const message = globalThis.Object.create(this.messagePrototype);
        message.markers = [];
        if (value !== undefined)
            reflectionMergePartial<AppMapMarkers>(this, message, value);
        return message;
    }
    internalBinaryRead(
        reader: IBinaryReader,
        length: number,
        options: BinaryReadOptions,
        target?: AppMapMarkers
    ): AppMapMarkers {
        const message = target ?? this.create(),
            end = reader.pos + length;
        while (reader.pos < end) {
            const [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated rustplus.AppMarker markers */ 1:
                    message.markers.push(
                        AppMarker.internalBinaryRead(
                            reader,
                            reader.uint32(),
                            options
                        )
                    );
                    break;
                default:
                    const u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(
                            `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
                        );
                    const d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(
                            this.typeName,
                            message,
                            fieldNo,
                            wireType,
                            d
                        );
            }
        }
        return message;
    }
    internalBinaryWrite(
        message: AppMapMarkers,
        writer: IBinaryWriter,
        options: BinaryWriteOptions
    ): IBinaryWriter {
        /* repeated rustplus.AppMarker markers = 1; */
        for (let i = 0; i < message.markers.length; i++)
            AppMarker.internalBinaryWrite(
                message.markers[i],
                writer.tag(1, WireType.LengthDelimited).fork(),
                options
            ).join();
        const u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(
                this.typeName,
                message,
                writer
            );
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message rustplus.AppMapMarkers
 */
export const AppMapMarkers = new AppMapMarkers$Type();